refactor: split the admin handlers by resource and share helpers
